dm_stub : fix MISRA.ETYPE.CATEGORY.DIFFERENT.2012
MISRA.ETYPE.CATEGORY.DIFFERENT.2012
The operands of an operator in which the usual arithmetic conversions
are performed do not have the same essential type category.
Use stronger typing to assign same category for each operand.

Fixes: SYSFW-6253

+++ lpm/lpm_serial_8250.h
@@ -3,7 +3,7 @@
  *
  * Minimal driver for UART access
  *
- * Copyright (C) 2021-2022, Texas Instruments Incorporated
+ * Copyright (C) 2021-2023, Texas Instruments Incorporated
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -83,12 +83,12 @@
 #define UART_16550_MBR                  0xB0
 
 /* Line Control Register bits */
-#define UART_16550_LCR_DLAB             (1 << 7)        /* Divisor Latch Access */
-#define UART_16550_LCR_SETB             (1 << 6)        /* Set BREAK Condition */
-#define UART_16550_LCR_SETP             (1 << 5)        /* Set Parity to LCR[4] */
-#define UART_16550_LCR_EVEN             (1 << 4)        /* Even Parity Format */
-#define UART_16550_LCR_PAR              (1 << 3)        /* Parity */
-#define UART_16550_LCR_STOP             (1 << 2)        /* Stop Bit */
+#define UART_16550_LCR_DLAB             (1U << 7U)      /* Divisor Latch Access */
+#define UART_16550_LCR_SETB             (1U << 6U)      /* Set BREAK Condition */
+#define UART_16550_LCR_SETP             (1U << 5U)      /* Set Parity to LCR[4] */
+#define UART_16550_LCR_EVEN             (1U << 4U)      /* Even Parity Format */
+#define UART_16550_LCR_PAR              (1U << 3U)      /* Parity */
+#define UART_16550_LCR_STOP             (1U << 2U)      /* Stop Bit */
 #define UART_16550_LCR_WORDSZ_5         0               /* Word Length of 5 */
 #define UART_16550_LCR_WORDSZ_6         1               /* Word Length of 6 */
 #define UART_16550_LCR_WORDSZ_7         2               /* Word Length of 7 */
@@ -109,8 +109,8 @@
 #define UART_16550_FCR_FIFOEN           (1 << 0)        /* Enable the Tx/Rx FIFO */
 
 /* Line Status Register bits */
-#define UART_16550_LSR_SR_E             (1 << 6)        /* Shift register empty */
-#define UART_16550_LSR_TX_FIFO_E        (1 << 5)        /* TX FIFO empty */
+#define UART_16550_LSR_SR_E             (1U << 6U)              /* Shift register empty */
+#define UART_16550_LSR_TX_FIFO_E        (1U << 5U)              /* TX FIFO empty */
 
 /* Modem Control Register bits */
 #define UART_16550_MCR_RTS              (1 << 0)        /* Request to Send */
@@ -130,7 +130,7 @@ extern const struct uart_16550_config soc_uart_16550_boot_config;
 extern const struct uart_16550_config soc_uart_16550_reinit_config;
 
 void lpm_serial_8250_init(const struct uart_16550_config *cfg);
-int lpm_console_tx(int data);
+int lpm_console_tx(u8 data);
 int lpm_puts(const char *str);
 
 void lpm_console_init(void);
